requirements.txt
torch>=2.0.0
torchvision
numpy
opencv-python
albumentations
pandas
seaborn
Pillow
scikit-learn
matplotlib
segmentation-models-pytorch
tqdm